Move SectionContainer from Sections component into each section
