+++ next/src/components/sections/CalculatorSection_Deprecated/CalculatorSection_Deprecated.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import MinimumCalculator from '@/src/components/sections/CalculatorSection_Deprecated/MinimumCalculator_Deprecated'
 import { CalculatorSectionFragment } from '@/src/services/graphql'
 
@@ -9,11 +10,13 @@ type CalculatorSectionProps = {
 
 const CalculatorSection = ({ section }: CalculatorSectionProps) => {
   return (
-    <MinimumCalculator
-      singleAdultValue={section.single_adult_value ?? 0}
-      anotherAdultValue={section.another_adult_value ?? 0}
-      childValue={section.child_value ?? 0}
-    />
+    <SectionContainer>
+      <MinimumCalculator
+        singleAdultValue={section.single_adult_value ?? 0}
+        anotherAdultValue={section.another_adult_value ?? 0}
+        childValue={section.child_value ?? 0}
+      />
+    </SectionContainer>
   )
 }
