+++ next/src/components/sections/DividerSection.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import Divider from '@/src/components/common/Divider/Divider'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { DividerSectionFragment } from '@/src/services/graphql'
 
 type DividerSectionProps = {
@@ -8,7 +9,11 @@ type DividerSectionProps = {
 }
 
 const DividerSection = ({ section }: DividerSectionProps) => {
-  return <Divider dividerStyle={section.style ?? undefined} />
+  return (
+    <SectionContainer>
+      <Divider dividerStyle={section.style ?? undefined} />
+    </SectionContainer>
+  )
 }
 
 export default DividerSection
