+++ next/src/components/sections/FaqsSection.tsx
@@ -2,6 +2,7 @@ import { Typography } from '@bratislava/component-library'
 import React from 'react'
 
 import FaqsGroup from '@/src/components/common/FaqsGroup/FaqsGroup'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { FaqsSectionFragment } from '@/src/services/graphql'
 import { isDefined } from '@/src/utils/isDefined'
 
@@ -15,16 +16,18 @@ const FaqsSection = ({ section }: Props) => {
   const filteredFaqs = faqs?.data.filter(isDefined) ?? []
 
   return (
-    <div className="flex flex-col gap-6 lg:gap-12">
-      {title || text ? (
-        <div className="col-span-1 flex flex-col gap-3 md:col-span-5">
-          {title ? <Typography variant="h2">{title}</Typography> : null}
-          {text ? <Typography variant="p-default">{text}</Typography> : null}
-        </div>
-      ) : null}
+    <SectionContainer>
+      <div className="flex flex-col gap-6 lg:gap-12">
+        {title || text ? (
+          <div className="col-span-1 flex flex-col gap-3 md:col-span-5">
+            {title ? <Typography variant="h2">{title}</Typography> : null}
+            {text ? <Typography variant="p-default">{text}</Typography> : null}
+          </div>
+        ) : null}
 
-      <FaqsGroup faqs={filteredFaqs} />
-    </div>
+        <FaqsGroup faqs={filteredFaqs} />
+      </div>
+    </SectionContainer>
   )
 }
