+++ next/src/components/sections/FileListSection.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import FileList from '@/src/components/common/FileList/FileList'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { FileListSectionFragment } from '@/src/services/graphql'
 import { isDefined } from '@/src/utils/isDefined'
 
@@ -10,11 +11,13 @@ type FileListSectionProps = {
 
 const FileListSection = ({ section }: FileListSectionProps) => {
   return (
-    <FileList
-      title={section.title}
-      text={section.text}
-      files={section.fileList?.filter(isDefined) ?? []}
-    />
+    <SectionContainer>
+      <FileList
+        title={section.title}
+        text={section.text}
+        files={section.fileList?.filter(isDefined) ?? []}
+      />
+    </SectionContainer>
   )
 }
