+++ next/src/components/sections/InbaArticlesListSection.tsx
@@ -6,6 +6,7 @@ import ArticleCard from '@/src/components/cards/ArticleCard'
 import { transformInbaArticleProps } from '@/src/components/cards/transformInbaArticleProps'
 import InbaArticlesFilter from '@/src/components/common/InbaArticlesFilter/InbaArticlesFilter'
 import Pagination from '@/src/components/common/Pagination/Pagination'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { InbaArticlesListSectionFragment } from '@/src/services/graphql'
 import { client } from '@/src/services/graphql/gql'
 import {
@@ -49,34 +50,36 @@ const InbaArticlesListSection = ({ section }: Props) => {
   }
 
   return (
-    <div className="flex flex-col gap-8">
-      <InbaArticlesFilter tags={tagData?.inbaTags?.data || []} onChange={handleTagFilterChange} />
-      {title || text ? (
-        <div className="flex flex-col gap-2">
-          {title && <Typography variant="h2">{title}</Typography>}
-          {text && <Typography variant="p-default">{text}</Typography>}
+    <SectionContainer>
+      <div className="flex flex-col gap-8">
+        <InbaArticlesFilter tags={tagData?.inbaTags?.data || []} onChange={handleTagFilterChange} />
+        {title || text ? (
+          <div className="flex flex-col gap-2">
+            {title && <Typography variant="h2">{title}</Typography>}
+            {text && <Typography variant="p-default">{text}</Typography>}
+          </div>
+        ) : null}
+        <div className="grid grid-cols-1 gap-8 sm:grid-cols-2 lg:grid-cols-3">
+          {data?.hits.map((card) => {
+            return card.attributes ? (
+              <ArticleCard
+                key={card.attributes.slug}
+                {...transformInbaArticleProps(card.attributes)}
+              />
+            ) : null
+          })}
         </div>
-      ) : null}
-      <div className="grid grid-cols-1 gap-8 sm:grid-cols-2 lg:grid-cols-3">
-        {data?.hits.map((card) => {
-          return card.attributes ? (
-            <ArticleCard
-              key={card.attributes.slug}
-              {...transformInbaArticleProps(card.attributes)}
-            />
-          ) : null
-        })}
-      </div>
 
-      {data?.estimatedTotalHits ? (
-        <Pagination
-          key={filters.search}
-          totalCount={Math.ceil(data.estimatedTotalHits / filters.pageSize)}
-          currentPage={filters.page}
-          onPageChange={handlePageChange}
-        />
-      ) : null}
-    </div>
+        {data?.estimatedTotalHits ? (
+          <Pagination
+            key={filters.search}
+            totalCount={Math.ceil(data.estimatedTotalHits / filters.pageSize)}
+            currentPage={filters.page}
+            onPageChange={handlePageChange}
+          />
+        ) : null}
+      </div>
+    </SectionContainer>
   )
 }
