+++ next/src/components/sections/NarrowTextSection.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 
 import NarrowText from '@/src/components/common/NarrowText/NarrowText'
 import Markdown from '@/src/components/formatting/Markdown/Markdown'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { NarrowTextSectionFragment } from '@/src/services/graphql'
 
 type NarrowTextSectionProps = {
@@ -14,9 +15,11 @@ const NarrowTextSection = ({ section }: NarrowTextSectionProps) => {
   }
 
   return (
-    <NarrowText align={section.align ?? undefined} width={section.width ?? undefined}>
-      <Markdown content={section.content} />
-    </NarrowText>
+    <SectionContainer>
+      <NarrowText align={section.align ?? undefined} width={section.width ?? undefined}>
+        <Markdown content={section.content} />
+      </NarrowText>
+    </SectionContainer>
   )
 }
