+++ next/src/components/sections/NumericalListSection.tsx
@@ -2,6 +2,7 @@
 import React from 'react'
 
 import NumericalList from '@/src/components/common/NumericalList_Deprecated/NumericalList_Deprecated'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { NumericalListSectionFragment } from '@/src/services/graphql'
 import { isPresent } from '@/src/utils/utils'
 
@@ -11,13 +12,15 @@ type NumericalListSectionProps = {
 
 const NumericalListSection = ({ section }: NumericalListSectionProps) => {
   return (
-    <NumericalList
-      title={section.title}
-      items={section.items?.filter(isPresent)}
-      buttonText={section.buttonText}
-      buttonLink={section.buttonLink}
-      variant={section.variant ?? undefined}
-    />
+    <SectionContainer>
+      <NumericalList
+        title={section.title}
+        items={section.items?.filter(isPresent)}
+        buttonText={section.buttonText}
+        buttonLink={section.buttonLink}
+        variant={section.variant ?? undefined}
+      />
+    </SectionContainer>
   )
 }
