+++ next/src/components/sections/OfficialBoardSection/OfficialBoardSection.tsx
@@ -1,9 +1,14 @@
 import React from 'react'
 
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import GlobalSearchSectionContent from '@/src/components/sections/SearchSection/GlobalSearchSectionContent'
 
 const OfficialBoardSection = () => {
-  return <GlobalSearchSectionContent variant="specific" searchOption="officialBoard" />
+  return (
+    <SectionContainer>
+      <GlobalSearchSectionContent variant="specific" searchOption="officialBoard" />
+    </SectionContainer>
+  )
 }
 
 export default OfficialBoardSection
