+++ next/src/components/sections/OrganizationalStructureSection.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import OrganizationalStructure from '@/src/components/common/OrganizationalStructure_Deprecated/OrganizationalStructure_Deprecated'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { OrganizationalStructureSectionFragment } from '@/src/services/graphql'
 
 type OrganizationalStructureSectionProps = {
@@ -8,7 +9,11 @@ type OrganizationalStructureSectionProps = {
 }
 
 const OrganizationalStructureSection = ({ section }: OrganizationalStructureSectionProps) => {
-  return <OrganizationalStructure title={section.title} />
+  return (
+    <SectionContainer>
+      <OrganizationalStructure title={section.title} />
+    </SectionContainer>
+  )
 }
 
 export default OrganizationalStructureSection
