+++ next/src/components/sections/RegulationsListSection.tsx
@@ -1,9 +1,14 @@
 import React from 'react'
 
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import GlobalSearchSectionContent from '@/src/components/sections/SearchSection/GlobalSearchSectionContent'
 
 const RegulationsListSection = () => {
-  return <GlobalSearchSectionContent variant="specific" searchOption="regulations" />
+  return (
+    <SectionContainer>
+      <GlobalSearchSectionContent variant="specific" searchOption="regulations" />
+    </SectionContainer>
+  )
 }
 
 export default RegulationsListSection
