+++ next/src/components/sections/VideosSection.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import { Videos } from '@/src/components/common/Videos_Deprecated/Videos_Deprecated'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { VideosSectionFragment } from '@/src/services/graphql'
 import { isPresent } from '@/src/utils/utils'
 
@@ -8,12 +9,14 @@ type VideosSectionProps = { section: VideosSectionFragment }
 
 const VideosSection = ({ section }: VideosSectionProps) => {
   return (
-    <Videos
-      id={section.id}
-      title={section.title}
-      subtitle={section.subtitle}
-      videos={section?.videos?.filter(isPresent) ?? null}
-    />
+    <SectionContainer>
+      <Videos
+        id={section.id}
+        title={section.title}
+        subtitle={section.subtitle}
+        videos={section?.videos?.filter(isPresent) ?? null}
+      />
+    </SectionContainer>
   )
 }
